# src/main/resources/swagger.json
{"swagger":"2.0","info":{"description":"API to record research for ORP","version":"0.1.0","title":"SuiviRecherches API","termsOfService":"Free to use","contact":{"name":"Mathieu Casse","url":"http://matfly.ch","email":"mc@matfly.ch"},"license":{"name":"API licence","url":"http://matfly.ch"}},"host":"localhost:8080","basePath":"/SuiviRecherches","tags":[{"name":"audit-rest-controller","description":"Audit Rest Controller"},{"name":"event-rest-controller","description":"Event Rest Controller"},{"name":"recherche-rest-controller","description":"Recherche Rest Controller"},{"name":"util-rest-controller","description":"Util Rest Controller"}],"paths":{"/rest/approcheMedia":{"get":{"tags":["util-rest-controller"],"summary":"Get All Possible Values for approcheMedia","operationId":"getApprocheMediaUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/assignationOrp":{"get":{"tags":["util-rest-controller"],"summary":"Get All Possible Values for assignationOrp","operationId":"getAssignationOrpUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/audit/entreprise/{id}":{"get":{"tags":["audit-rest-controller"],"summary":"getEntrepriseAudit","operationId":"getEntrepriseAuditUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/EntrepriseAuditDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/audit/evenement/{id}":{"get":{"tags":["audit-rest-controller"],"summary":"getEvenementAudit","operationId":"getEvenementAuditUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/EvenementAuditDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/audit/personne/{id}":{"get":{"tags":["audit-rest-controller"],"summary":"getPersonneAudit","operationId":"getPersonneAuditUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/PersonneAuditDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/audit/recherche/{id}":{"get":{"tags":["audit-rest-controller"],"summary":"getRechercheAudit","operationId":"getRechercheAuditUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/RechercheAuditDto"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/event":{"post":{"tags":["event-rest-controller"],"summary":"add an event to a recherche","description":"add an event to recherche","operationId":"addEventUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"evenement","description":"evenement","required":true,"schema":{"$ref":"#/definitions/Evenement"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Evenement"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["event-rest-controller"],"summary":"update an event","description":"update an event from an Event","operationId":"updateEventUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"evenement","description":"evenement","required":true,"schema":{"$ref":"#/definitions/Evenement"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AngularRechercheDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/event/byRechercheId/{id}":{"get":{"tags":["event-rest-controller"],"summary":"Get All Events of a recherche","description":"Return a list of AngularRechercheDTO","operationId":"getEventsByRechercheUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/event/{id}":{"delete":{"tags":["event-rest-controller"],"summary":"delete  an event","description":"delete an event from an id","operationId":"delEvenementUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AngularRechercheDto"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/rest/recherche":{"post":{"tags":["recherche-rest-controller"],"summary":"add a recherche","description":"add a recherche from an AngularRechercheDTO","operationId":"addRechercheUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"aRecherche","description":"aRecherche","required":true,"schema":{"$ref":"#/definitions/AngularRechercheDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AngularRechercheDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["recherche-rest-controller"],"summary":"update a recherche","description":"update a recherche from an AngularRechercheDTO","operationId":"updateRechercheUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"aRecherche","description":"aRecherche","required":true,"schema":{"$ref":"#/definitions/AngularRechercheDto"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AngularRechercheDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/recherche/{id}":{"delete":{"tags":["recherche-rest-controller"],"summary":"delete a recherche","description":"delete a recherche from an id","operationId":"delRechercheUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AngularRechercheDto"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/rest/rechercheStatut":{"get":{"tags":["util-rest-controller"],"summary":"Get All Possible Values for rechercheStatut","operationId":"getRechercheStatutUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/recherches":{"get":{"tags":["recherche-rest-controller"],"summary":"Get All Recherches","description":"Return a list of AngularRechercheDTO","operationId":"getRecherchesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/recherches/{email}":{"get":{"tags":["recherche-rest-controller"],"summary":"Get All Recherches By User email","description":"Return for a User the list of AngularRechercheDTO ","operationId":"getRecherchesByUserEmailUsingGET","produces":["*/*"],"parameters":[{"name":"email","in":"path","description":"email","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/rest/tauxActivite":{"get":{"tags":["util-rest-controller"],"summary":"Get All Possible Values for tauxActivite","operationId":"getTauxActiviteUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"AngularRechercheDto":{"type":"object","properties":{"approcheMedia":{"type":"string"},"assignationORP":{"type":"string"},"contactEmailF":{"type":"string"},"contactEmailS":{"type":"string"},"contactNomF":{"type":"string"},"contactNomS":{"type":"string"},"contactPrenomF":{"type":"string"},"contactPrenomS":{"type":"string"},"contactTelephoneF":{"type":"string"},"contactTelephoneS":{"type":"string"},"dateContact":{"type":"string"},"entrepriseF":{"type":"string"},"entrepriseS":{"type":"string"},"entrepriseTelF":{"type":"string"},"entrepriseTelS":{"type":"string"},"id":{"type":"integer","format":"int64"},"poste":{"type":"string"},"statut":{"type":"string"},"tauxActivite":{"type":"integer","format":"int32"}},"title":"AngularRechercheDto"},"EntrepriseAuditDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"rev":{"type":"integer","format":"int32"},"revtstmp":{"type":"string"},"revtype":{"type":"integer","format":"int32"},"telephone":{"type":"string"}},"title":"EntrepriseAuditDto"},"Evenement":{"type":"object","properties":{"eventDate":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"rechercheId":{"type":"integer","format":"int64"},"texte":{"type":"string"},"titre":{"type":"string"}},"title":"Evenement"},"EvenementAuditDto":{"type":"object","properties":{"eventDate":{"type":"string"},"id":{"type":"integer","format":"int64"},"recherche_id":{"type":"integer","format":"int64"},"rev":{"type":"integer","format":"int32"},"revtstmp":{"type":"string"},"revtype":{"type":"integer","format":"int32"},"texte":{"type":"string"},"titre":{"type":"string"}},"title":"EvenementAuditDto"},"PersonneAuditDto":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"nom":{"type":"string"},"prenom":{"type":"string"},"rev":{"type":"integer","format":"int32"},"revtstmp":{"type":"string"},"revtype":{"type":"integer","format":"int32"},"telephone":{"type":"string"}},"title":"PersonneAuditDto"},"RechercheAuditDto":{"type":"object","properties":{"approchemedia":{"type":"string"},"assignationORP":{"type":"string"},"contactdate":{"type":"string"},"entrepriseF_id":{"type":"integer","format":"int64"},"entrepriseS_id":{"type":"integer","format":"int64"},"id":{"type":"integer","format":"int64"},"personneF_id":{"type":"integer","format":"int64"},"personneS_id":{"type":"integer","format":"int64"},"poste":{"type":"string"},"rev":{"type":"integer","format":"int32"},"revtstmp":{"type":"string"},"revtype":{"type":"integer","format":"int32"},"statut":{"type":"string"},"tauxactivite":{"type":"string"}},"title":"RechercheAuditDto"}}}